CODETOOLS-7902928: Improve support for async-profiler 2.x
Allow multiple events to be captured simultaneously in this version,
provided that JFR is chosen as output format.

Delegate output file writing to the async-profiler. This is required
in 2.x for the JFR output but is supported in both versions. The file
path must be provided when starting the profiler with JFR output, so
we need to create the per-trial output directory in the first
`beforeIteration`.

Avoid the character '%' in the generated directory name as this
is interpreted by async-profiler as part %p or %t placeholder.