Allow forcing the use of compiler hints using a system property
Using `-Djmh.compilerhints.mode=FORCE_ON` will force the use of compiler
hints. `-Djmh.compilerhints.mode=FORCE_OFF` will prevent the use of compiler
hints. When the system property is not set, or when it is set to `AUTO`,
the current beahviour based on vm name and version will be used.