README Preferred Usage Improvement
Attempting to execute the "mvn archetype:generate" command specified in Step 1 of Preferred Usage in the README will result in a maven execution error.

In my case I naively interpreted the following text to mean that I could create a sub project within the test folder of my already existing java project. "test" is the default test folder!

This resulted in 10-15 minutes of trying to figure out if I needed to install something new to use mvn archetype - as I have not used it before.

This change might make it easier for users to get set up with JMH project templates.