7903560: JMH: async profiler reports only the last fork result
Without this change, async-profiler would write the same result in each fork, overwriting the previous result.

This patch introduces a new "FileResult" class that collects the output paths of each benchmark. In afterTrial, the output files are copied to a new path that includes the pid (fork number would be an alternative but it's not as easily available).

Reviewed-by: shade